build: add back in resources/_gen for netlify
This might happen if you’re using image processing. Is that the case?

If it is, it’s useful to commit the resources/_gen folder from the local to the repo. Doing so will help Hugo not have to process images again, and thus the timeout won’t be triggered.

A loop can also happen when you’re trying to access non-shortcode variables in a shortcode using some hacky way.